refactor: move CLI path helpers from shared utils
Relocated collectAncestorDirectories and isPathInside from src/shared/utils/path.js to src/cli/lib/path-utils.js because the CLI is their only consumer and they enforce CLI traversal rules. Updated CLI imports/tests to use the new module so shared path-utils only exposes cross-package POSIX conversion helpers.
